Drawerからall-twoshotsページに中身を移動
