+++ src/components/MyDrawer.tsx
@@ -1,4 +1,4 @@
-import { ExternalLinkIcon, HamburgerIcon } from "@chakra-ui/icons";
+import { HamburgerIcon } from "@chakra-ui/icons";
 import {
   Button,
   Drawer,
@@ -9,20 +9,11 @@ import {
   DrawerHeader,
   DrawerOverlay,
   IconButton,
-  Link,
-  ListItem,
-  OrderedList,
   useDisclosure,
 } from "@chakra-ui/react";
-import { useState } from "react";
-import { HiOutlineSwitchVertical } from "react-icons/hi";
-import rawJson from "../assets/raw.json";
-
-const reversedAllTwoshots = rawJson.slice().reverse();
 
 const MyDrawer = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const [reversed, setReversed] = useState(false);
 
   return (
     <>
@@ -36,30 +27,8 @@ const MyDrawer = () => {
         <DrawerContent>
           <DrawerCloseButton />
           <DrawerHeader>ALL TWOSHOTs</DrawerHeader>
-          <DrawerBody>
-            <OrderedList
-              spacing={4}
-              start={reversed ? rawJson.length - 1 : 0}
-              reversed={reversed}
-            >
-              {(reversed ? reversedAllTwoshots : rawJson).map(
-                ({ title, videoId }, i) => (
-                  <ListItem key={i}>
-                    <Link href={`https://youtu.be/${videoId}`} isExternal>
-                      {title}
-                      <ExternalLinkIcon color="green.500" />
-                    </Link>
-                  </ListItem>
-                )
-              )}
-            </OrderedList>
-          </DrawerBody>
+          <DrawerBody></DrawerBody>
           <DrawerFooter>
-            <IconButton
-              aria-label="Reverse"
-              icon={<HiOutlineSwitchVertical />}
-              onClick={() => setReversed(!reversed)}
-            />
             <Button ml={1} onClick={onClose}>
               Close
             </Button>
